src/main/java/roomescape/service/MemberService.java
package roomescape.service;

import java.util.List;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import roomescape.controller.dto.FindMemberResponse;
import roomescape.domain.member.Member;
import roomescape.global.exception.RoomescapeException;
import roomescape.repository.MemberRepository;

@Service
public class MemberService {

    private final MemberRepository memberRepository;

    public MemberService(MemberRepository memberRepository) {
        this.memberRepository = memberRepository;
    }

    @Transactional(readOnly = true)
    public List<FindMemberResponse> findAll() {
        List<Member> members = memberRepository.findAll();
        return members.stream()
            .map(FindMemberResponse::from)
            .toList();
    }

    @Transactional(readOnly = true)
    public Member findById(Long memberId) {
        return memberRepository.findById(memberId)
            .orElseThrow(() -> new RoomescapeException("입력한 사용자 ID에 해당하는 데이터가 존재하지 않습니다."));
    }
}
